+++ crates/rust-client/src/store/mod.rs
@@ -22,6 +22,7 @@
 
 use alloc::boxed::Box;
 use alloc::collections::{BTreeMap, BTreeSet};
+use alloc::string::String;
 use alloc::vec::Vec;
 use core::fmt::Debug;
 
@@ -276,6 +277,21 @@ pub trait Store: Send + Sync {
     ///
     /// Returns a `StoreError::AccountDataNotFound` if there is no account for the provided ID.
     async fn update_account(&self, new_account_state: &Account) -> Result<(), StoreError>;
+
+    // SETTINGS
+    // --------------------------------------------------------------------------------------------
+
+    /// Adds a value to the `settings` table.
+    async fn set_setting(&self, key: String, value: Vec<u8>) -> Result<(), StoreError>;
+
+    /// Retrieves a value from the `settings` table.
+    async fn get_setting(&self, key: String) -> Result<Option<Vec<u8>>, StoreError>;
+
+    /// Deletes a value from the `settings` table.
+    async fn remove_setting(&self, key: String) -> Result<(), StoreError>;
+
+    /// Returns all the keys from the `settings` table.
+    async fn list_setting_keys(&self) -> Result<Vec<String>, StoreError>;
 
     // SYNC
     // --------------------------------------------------------------------------------------------