+++ crates/web-client/src/rpc_client/mod.rs
@@ -5,10 +5,10 @@
 use alloc::sync::Arc;
 use alloc::vec::Vec;
 
+use miden_client::note::NoteId as NativeNoteId;
 use miden_client::rpc::domain::note::FetchedNote as NativeFetchedNote;
 use miden_client::rpc::{NodeRpcClient, TonicRpcClient};
-use miden_objects::note::NoteId as NativeNoteId;
-use miden_objects::transaction::InputNote as NativeInputNote;
+use miden_client::transaction::InputNote as NativeInputNote;
 use note::FetchedNote;
 use wasm_bindgen::prelude::*;